Make request_timeout_seconds a whole number of seconds
qdrant-client takes its timeout as an int and rounds a fraction up, so a
fractional value was honored by pymilvus and silently changed for Qdrant.
An int is honored exactly by both, and matches max_retry_interval_seconds
on the embedder and language model configurations.

Folded from #1630 into #1631: the whole seconds reach the configurations and
tests that also name a collection_registry.

(cherry picked from commit 5e9f36385b39ff645830ec60f4af4b6f544e2736)
(cherry picked from commit 874ec470738810654068c845a8b5642716850ae3)

+++ packages/server/src/memmachine_server/common/vector_store/milvus_vector_store.py
@@ -180,7 +180,7 @@ class MilvusVectorStoreCollection(VectorStoreCollection):
         config: VectorStoreCollectionConfig,
         tracker: OperationTracker,
         is_live: Callable[[UUID], Awaitable[bool]],
-        request_timeout_seconds: float,
+        request_timeout_seconds: int,
     ) -> None:
         """Initialize with a Milvus client and the incarnation the handle is bound to."""
         self._client = client
@@ -482,7 +482,7 @@ class MilvusVectorStoreParams(BaseModel):
             process, uses this registry, and no store on another
             deployment does.
         consistency_level (str): Collection consistency level for newly created collections.
-        request_timeout_seconds (float): Seconds any request to Milvus may take.
+        request_timeout_seconds (int): Seconds any request to Milvus may take.
         metrics_factory (MetricsFactory | None): Metrics factory for collecting usage metrics.
     """
 
@@ -498,7 +498,7 @@ class MilvusVectorStoreParams(BaseModel):
         default="Session",
         description="Milvus consistency level for newly created collections",
     )
-    request_timeout_seconds: float = Field(
+    request_timeout_seconds: int = Field(
         ..., gt=0, description="Seconds any request to Milvus may take"
     )
     metrics_factory: InstanceOf[MetricsFactory] | None = Field(
